backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
requests>=2.31
django-filter>=23.5
drf-spectacular>=0.27
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
channels>=4.0
channels-redis>=4.2
daphne>=4.0
django-redis>=5.4
